+++ typescript/packages/mechanisms/evm/src/exact/facilitator/scheme.ts
@@ -5,24 +5,15 @@ import {
   SettleResponse,
   VerifyResponse,
 } from "@x402/core/types";
-import { getAddress, Hex, isAddressEqual, parseErc6492Signature, parseSignature } from "viem";
-import {
-  authorizationTypes,
-  eip3009ABI,
-  permit2WitnessTypes,
-  PERMIT2_ADDRESS,
-  x402ExactPermit2ProxyAddress,
-  x402ExactPermit2ProxyABI,
-} from "../../constants";
-import { EIP2612_GAS_SPONSORING_EXTENSION } from "../client/permit2";
 import { FacilitatorEvmSigner } from "../../signer";
 import {
-  EIP2612PermitParams,
   ExactEIP3009Payload,
   ExactEvmPayloadV2,
   ExactPermit2Payload,
   isPermit2Payload,
 } from "../../types";
+import { verifyEIP3009, settleEIP3009 } from "./eip3009";
+import { verifyPermit2, settlePermit2 } from "./permit2";
 
 export interface ExactEvmSchemeConfig {
   /**
@@ -36,6 +27,10 @@ export interface ExactEvmSchemeConfig {
 
 /**
  * EVM facilitator implementation for the Exact payment scheme.
+ * Supports both EIP-3009 (transferWithAuthorization) and Permit2 flows.
+ *
+ * Routes to the appropriate verification and settlement method based on
+ * the payload type (EIP-3009 or Permit2).
  */
 export class ExactEvmScheme implements SchemeNetworkFacilitator {
   readonly scheme = "exact";
@@ -81,6 +76,7 @@ export class ExactEvmScheme implements SchemeNetworkFacilitator {
 
   /**
    * Verifies a payment payload.
+   * Routes to EIP-3009 or Permit2 verification based on payload type.
    *
    * @param payload - The payment payload to verify
    * @param requirements - The payment requirements
@@ -93,192 +89,15 @@ export class ExactEvmScheme implements SchemeNetworkFacilitator {
     const rawPayload = payload.payload as ExactEvmPayloadV2;
 
     if (isPermit2Payload(rawPayload)) {
-      return this.verifyPermit2(payload, requirements, rawPayload);
+      return verifyPermit2(this.signer, payload, requirements, rawPayload);
     }
 
-    const exactEvmPayload: ExactEIP3009Payload = rawPayload;
-
-    // Verify scheme matches
-    if (payload.accepted.scheme !== "exact" || requirements.scheme !== "exact") {
-      return {
-        isValid: false,
-        invalidReason: "unsupported_scheme",
-        payer: exactEvmPayload.authorization.from,
-      };
-    }
-
-    // Get chain configuration
-    if (!requirements.extra?.name || !requirements.extra?.version) {
-      return {
-        isValid: false,
-        invalidReason: "missing_eip712_domain",
-        payer: exactEvmPayload.authorization.from,
-      };
-    }
-
-    const { name, version } = requirements.extra;
-    const erc20Address = getAddress(requirements.asset);
-
-    // Verify network matches
-    if (payload.accepted.network !== requirements.network) {
-      return {
-        isValid: false,
-        invalidReason: "network_mismatch",
-        payer: exactEvmPayload.authorization.from,
-      };
-    }
-
-    // Build typed data for signature verification
-    const permitTypedData = {
-      types: authorizationTypes,
-      primaryType: "TransferWithAuthorization" as const,
-      domain: {
-        name,
-        version,
-        chainId: parseInt(requirements.network.split(":")[1]),
-        verifyingContract: erc20Address,
-      },
-      message: {
-        from: exactEvmPayload.authorization.from,
-        to: exactEvmPayload.authorization.to,
-        value: BigInt(exactEvmPayload.authorization.value),
-        validAfter: BigInt(exactEvmPayload.authorization.validAfter),
-        validBefore: BigInt(exactEvmPayload.authorization.validBefore),
-        nonce: exactEvmPayload.authorization.nonce,
-      },
-    };
-
-    // Verify signature
-    try {
-      const recoveredAddress = await this.signer.verifyTypedData({
-        address: exactEvmPayload.authorization.from,
-        ...permitTypedData,
-        signature: exactEvmPayload.signature!,
-      });
-
-      if (!recoveredAddress) {
-        return {
-          isValid: false,
-          invalidReason: "invalid_exact_evm_payload_signature",
-          payer: exactEvmPayload.authorization.from,
-        };
-      }
-    } catch {
-      // Signature verification failed - could be an undeployed smart wallet
-      // Check if smart wallet is deployed
-      const signature = exactEvmPayload.signature!;
-      const signatureLength = signature.startsWith("0x") ? signature.length - 2 : signature.length;
-      const isSmartWallet = signatureLength > 130; // 65 bytes = 130 hex chars for EOA
-
-      if (isSmartWallet) {
-        const payerAddress = exactEvmPayload.authorization.from;
-        const bytecode = await this.signer.getCode({ address: payerAddress });
-
-        if (!bytecode || bytecode === "0x") {
-          // Wallet is not deployed. Check if it's EIP-6492 with deployment info.
-          // EIP-6492 signatures contain factory address and calldata needed for deployment.
-          // Non-EIP-6492 undeployed wallets cannot succeed (no way to deploy them).
-          const erc6492Data = parseErc6492Signature(signature);
-          const hasDeploymentInfo =
-            erc6492Data.address &&
-            erc6492Data.data &&
-            !isAddressEqual(erc6492Data.address, "0x0000000000000000000000000000000000000000");
-
-          if (!hasDeploymentInfo) {
-            // Non-EIP-6492 undeployed smart wallet - will always fail at settlement
-            // since EIP-3009 requires on-chain EIP-1271 validation
-            return {
-              isValid: false,
-              invalidReason: "invalid_exact_evm_payload_undeployed_smart_wallet",
-              payer: payerAddress,
-            };
-          }
-          // EIP-6492 signature with deployment info - allow through
-          // Facilitators with sponsored deployment support can handle this in settle()
-        } else {
-          // Wallet is deployed but signature still failed - invalid signature
-          return {
-            isValid: false,
-            invalidReason: "invalid_exact_evm_payload_signature",
-            payer: exactEvmPayload.authorization.from,
-          };
-        }
-      } else {
-        // EOA signature failed
-        return {
-          isValid: false,
-          invalidReason: "invalid_exact_evm_payload_signature",
-          payer: exactEvmPayload.authorization.from,
-        };
-      }
-    }
-
-    // Verify payment recipient matches
-    if (getAddress(exactEvmPayload.authorization.to) !== getAddress(requirements.payTo)) {
-      return {
-        isValid: false,
-        invalidReason: "invalid_exact_evm_payload_recipient_mismatch",
-        payer: exactEvmPayload.authorization.from,
-      };
-    }
-
-    // Verify validBefore is in the future (with 6 second buffer for block time)
-    const now = Math.floor(Date.now() / 1000);
-    if (BigInt(exactEvmPayload.authorization.validBefore) < BigInt(now + 6)) {
-      return {
-        isValid: false,
-        invalidReason: "invalid_exact_evm_payload_authorization_valid_before",
-        payer: exactEvmPayload.authorization.from,
-      };
-    }
-
-    // Verify validAfter is not in the future
-    if (BigInt(exactEvmPayload.authorization.validAfter) > BigInt(now)) {
-      return {
-        isValid: false,
-        invalidReason: "invalid_exact_evm_payload_authorization_valid_after",
-        payer: exactEvmPayload.authorization.from,
-      };
-    }
-
-    // Check balance
-    try {
-      const balance = (await this.signer.readContract({
-        address: erc20Address,
-        abi: eip3009ABI,
-        functionName: "balanceOf",
-        args: [exactEvmPayload.authorization.from],
-      })) as bigint;
-
-      if (BigInt(balance) < BigInt(requirements.amount)) {
-        return {
-          isValid: false,
-          invalidReason: "insufficient_funds",
-          payer: exactEvmPayload.authorization.from,
-        };
-      }
-    } catch {
-      // If we can't check balance, continue with other validations
-    }
-
-    // Verify amount is sufficient
-    if (BigInt(exactEvmPayload.authorization.value) < BigInt(requirements.amount)) {
-      return {
-        isValid: false,
-        invalidReason: "invalid_exact_evm_payload_authorization_value",
-        payer: exactEvmPayload.authorization.from,
-      };
-    }
-
-    return {
-      isValid: true,
-      invalidReason: undefined,
-      payer: exactEvmPayload.authorization.from,
-    };
+    return verifyEIP3009(this.signer, payload, requirements, rawPayload as ExactEIP3009Payload);
   }
 
   /**
    * Settles a payment by executing the transfer.
+   * Routes to EIP-3009 or Permit2 settlement based on payload type.
    *
    * @param payload - The payment payload to settle
    * @param requirements - The payment requirements
@@ -291,427 +110,23 @@ export class ExactEvmScheme implements SchemeNetworkFacilitator {
     const rawPayload = payload.payload as ExactEvmPayloadV2;
 
     if (isPermit2Payload(rawPayload)) {
-      return this.settlePermit2(payload, requirements, rawPayload);
+      return settlePermit2(
+        this.signer,
+        payload,
+        requirements,
+        rawPayload as ExactPermit2Payload,
+        () => verifyPermit2(this.signer, payload, requirements, rawPayload),
+      );
     }
 
-    const exactEvmPayload: ExactEIP3009Payload = rawPayload;
-
-    // Re-verify before settling
-    const valid = await this.verify(payload, requirements);
-    if (!valid.isValid) {
-      return {
-        success: false,
-        network: payload.accepted.network,
-        transaction: "",
-        errorReason: valid.invalidReason ?? "invalid_scheme",
-        payer: exactEvmPayload.authorization.from,
-      };
-    }
-
-    try {
-      // Parse ERC-6492 signature if applicable
-      const parseResult = parseErc6492Signature(exactEvmPayload.signature!);
-      const { signature, address: factoryAddress, data: factoryCalldata } = parseResult;
-
-      // Deploy ERC-4337 smart wallet via EIP-6492 if configured and needed
-      if (
-        this.config.deployERC4337WithEIP6492 &&
-        factoryAddress &&
-        factoryCalldata &&
-        !isAddressEqual(factoryAddress, "0x0000000000000000000000000000000000000000")
-      ) {
-        // Check if smart wallet is already deployed
-        const payerAddress = exactEvmPayload.authorization.from;
-        const bytecode = await this.signer.getCode({ address: payerAddress });
-
-        if (!bytecode || bytecode === "0x") {
-          // Wallet not deployed - attempt deployment
-          try {
-            console.log(`Deploying ERC-4337 smart wallet for ${payerAddress} via EIP-6492`);
-
-            // Send the factory calldata directly as a transaction
-            // The factoryCalldata already contains the complete encoded function call
-            const deployTx = await this.signer.sendTransaction({
-              to: factoryAddress as Hex,
-              data: factoryCalldata as Hex,
-            });
-
-            // Wait for deployment transaction
-            await this.signer.waitForTransactionReceipt({ hash: deployTx });
-            console.log(`Successfully deployed smart wallet for ${payerAddress}`);
-          } catch (deployError) {
-            console.error("Smart wallet deployment failed:", deployError);
-            // Deployment failed - cannot proceed
-            throw deployError;
-          }
-        } else {
-          console.log(`Smart wallet for ${payerAddress} already deployed, skipping deployment`);
-        }
-      }
-
-      // Determine if this is an ECDSA signature (EOA) or smart wallet signature
-      // ECDSA signatures are exactly 65 bytes (130 hex chars without 0x)
-      const signatureLength = signature.startsWith("0x") ? signature.length - 2 : signature.length;
-      const isECDSA = signatureLength === 130;
-
-      let tx: Hex;
-      if (isECDSA) {
-        // For EOA wallets, parse signature into v, r, s and use that overload
-        const parsedSig = parseSignature(signature);
-
-        tx = await this.signer.writeContract({
-          address: getAddress(requirements.asset),
-          abi: eip3009ABI,
-          functionName: "transferWithAuthorization",
-          args: [
-            getAddress(exactEvmPayload.authorization.from),
-            getAddress(exactEvmPayload.authorization.to),
-            BigInt(exactEvmPayload.authorization.value),
-            BigInt(exactEvmPayload.authorization.validAfter),
-            BigInt(exactEvmPayload.authorization.validBefore),
-            exactEvmPayload.authorization.nonce,
-            (parsedSig.v as number | undefined) || parsedSig.yParity,
-            parsedSig.r,
-            parsedSig.s,
-          ],
-        });
-      } else {
-        // For smart wallets, use the bytes signature overload
-        // The signature contains WebAuthn/P256 or other ERC-1271 compatible signature data
-        tx = await this.signer.writeContract({
-          address: getAddress(requirements.asset),
-          abi: eip3009ABI,
-          functionName: "transferWithAuthorization",
-          args: [
-            getAddress(exactEvmPayload.authorization.from),
-            getAddress(exactEvmPayload.authorization.to),
-            BigInt(exactEvmPayload.authorization.value),
-            BigInt(exactEvmPayload.authorization.validAfter),
-            BigInt(exactEvmPayload.authorization.validBefore),
-            exactEvmPayload.authorization.nonce,
-            signature,
-          ],
-        });
-      }
-
-      // Wait for transaction confirmation
-      const receipt = await this.signer.waitForTransactionReceipt({ hash: tx });
-
-      if (receipt.status !== "success") {
-        return {
-          success: false,
-          errorReason: "invalid_transaction_state",
-          transaction: tx,
-          network: payload.accepted.network,
-          payer: exactEvmPayload.authorization.from,
-        };
-      }
-
-      return {
-        success: true,
-        transaction: tx,
-        network: payload.accepted.network,
-        payer: exactEvmPayload.authorization.from,
-      };
-    } catch (error) {
-      console.error("Failed to settle transaction:", error);
-      return {
-        success: false,
-        errorReason: "transaction_failed",
-        transaction: "",
-        network: payload.accepted.network,
-        payer: exactEvmPayload.authorization.from,
-      };
-    }
-  }
-
-  /**
-   * Verifies a Permit2 payment payload.
-   *
-   * @param payload - The payment payload to verify
-   * @param requirements - The payment requirements
-   * @param permit2Payload - The Permit2-specific payload data
-   * @returns Promise resolving to verification response
-   */
-  private async verifyPermit2(
-    payload: PaymentPayload,
-    requirements: PaymentRequirements,
-    permit2Payload: ExactPermit2Payload,
-  ): Promise<VerifyResponse> {
-    const payer = permit2Payload.permit2Authorization.from;
-
-    if (payload.accepted.scheme !== "exact" || requirements.scheme !== "exact") {
-      return {
-        isValid: false,
-        invalidReason: "unsupported_scheme",
-        payer,
-      };
-    }
-
-    if (payload.accepted.network !== requirements.network) {
-      return {
-        isValid: false,
-        invalidReason: "network_mismatch",
-        payer,
-      };
-    }
-
-    const chainId = parseInt(requirements.network.split(":")[1]);
-
-    if (
-      getAddress(permit2Payload.permit2Authorization.spender) !==
-      getAddress(x402ExactPermit2ProxyAddress)
-    ) {
-      return {
-        isValid: false,
-        invalidReason: "invalid_permit2_spender",
-        payer,
-      };
-    }
-
-    if (
-      getAddress(permit2Payload.permit2Authorization.witness.to) !== getAddress(requirements.payTo)
-    ) {
-      return {
-        isValid: false,
-        invalidReason: "invalid_permit2_witness_recipient",
-        payer,
-      };
-    }
-
-    if (
-      getAddress(permit2Payload.permit2Authorization.permitted.token) !==
-      getAddress(requirements.asset)
-    ) {
-      return {
-        isValid: false,
-        invalidReason: "invalid_permit2_token",
-        payer,
-      };
-    }
-
-    if (
-      BigInt(permit2Payload.permit2Authorization.permitted.amount) < BigInt(requirements.amount)
-    ) {
-      return {
-        isValid: false,
-        invalidReason: "invalid_permit2_amount",
-        payer,
-      };
-    }
-
-    const now = Math.floor(Date.now() / 1000);
-
-    if (BigInt(permit2Payload.permit2Authorization.deadline) < BigInt(now + 6)) {
-      return {
-        isValid: false,
-        invalidReason: "invalid_permit2_deadline",
-        payer,
-      };
-    }
-
-    if (BigInt(permit2Payload.permit2Authorization.witness.validBefore) < BigInt(now + 6)) {
-      return {
-        isValid: false,
-        invalidReason: "invalid_permit2_witness_valid_before",
-        payer,
-      };
-    }
-
-    if (BigInt(permit2Payload.permit2Authorization.witness.validAfter) > BigInt(now)) {
-      return {
-        isValid: false,
-        invalidReason: "invalid_permit2_witness_valid_after",
-        payer,
-      };
-    }
-
-    const permit2TypedData = {
-      types: permit2WitnessTypes,
-      primaryType: "PermitWitnessTransferFrom" as const,
-      domain: {
-        name: "Permit2",
-        chainId,
-        verifyingContract: PERMIT2_ADDRESS,
-      },
-      message: {
-        permitted: {
-          token: getAddress(permit2Payload.permit2Authorization.permitted.token),
-          amount: BigInt(permit2Payload.permit2Authorization.permitted.amount),
-        },
-        spender: getAddress(permit2Payload.permit2Authorization.spender),
-        nonce: BigInt(permit2Payload.permit2Authorization.nonce),
-        deadline: BigInt(permit2Payload.permit2Authorization.deadline),
-        witness: {
-          extra: permit2Payload.permit2Authorization.witness.extra,
-          to: getAddress(permit2Payload.permit2Authorization.witness.to),
-          validAfter: BigInt(permit2Payload.permit2Authorization.witness.validAfter),
-          validBefore: BigInt(permit2Payload.permit2Authorization.witness.validBefore),
-        },
-      },
-    };
-
-    try {
-      const recoveredAddress = await this.signer.verifyTypedData({
-        address: payer,
-        ...permit2TypedData,
-        signature: permit2Payload.signature,
-      });
-
-      if (!recoveredAddress) {
-        return {
-          isValid: false,
-          invalidReason: "invalid_permit2_signature",
-          payer,
-        };
-      }
-    } catch {
-      return {
-        isValid: false,
-        invalidReason: "invalid_permit2_signature",
-        payer,
-      };
-    }
-
-    try {
-      const balance = (await this.signer.readContract({
-        address: getAddress(requirements.asset),
-        abi: eip3009ABI,
-        functionName: "balanceOf",
-        args: [payer],
-      })) as bigint;
-
-      if (balance < BigInt(requirements.amount)) {
-        return {
-          isValid: false,
-          invalidReason: "insufficient_funds",
-          payer,
-        };
-      }
-    } catch {
-      // If we can't check balance, continue with other validations
-    }
-
-    return {
-      isValid: true,
-      invalidReason: undefined,
-      payer,
-    };
-  }
-
-  /**
-   * Settles a Permit2 payment by executing the transfer.
-   *
-   * @param payload - The payment payload to settle
-   * @param requirements - The payment requirements
-   * @param permit2Payload - The Permit2-specific payload data
-   * @returns Promise resolving to settlement response
-   */
-  private async settlePermit2(
-    payload: PaymentPayload,
-    requirements: PaymentRequirements,
-    permit2Payload: ExactPermit2Payload,
-  ): Promise<SettleResponse> {
-    const payer = permit2Payload.permit2Authorization.from;
-
-    const valid = await this.verifyPermit2(payload, requirements, permit2Payload);
-    if (!valid.isValid) {
-      return {
-        success: false,
-        network: payload.accepted.network,
-        transaction: "",
-        errorReason: valid.invalidReason ?? "invalid_payload",
-        payer,
-      };
-    }
-
-    try {
-      const permit = {
-        permitted: {
-          token: getAddress(permit2Payload.permit2Authorization.permitted.token),
-          amount: BigInt(permit2Payload.permit2Authorization.permitted.amount),
-        },
-        nonce: BigInt(permit2Payload.permit2Authorization.nonce),
-        deadline: BigInt(permit2Payload.permit2Authorization.deadline),
-      };
-
-      const witness = {
-        to: getAddress(permit2Payload.permit2Authorization.witness.to),
-        validAfter: BigInt(permit2Payload.permit2Authorization.witness.validAfter),
-        validBefore: BigInt(permit2Payload.permit2Authorization.witness.validBefore),
-        extra: permit2Payload.permit2Authorization.witness.extra,
-      };
-
-      const eip2612Extension = payload.extensions?.[EIP2612_GAS_SPONSORING_EXTENSION] as
-        | { permit: EIP2612PermitParams }
-        | undefined;
-
-      let tx: Hex;
-
-      if (eip2612Extension) {
-        const permit2612 = {
-          value: BigInt(eip2612Extension.permit.value),
-          deadline: BigInt(eip2612Extension.permit.deadline),
-          r: eip2612Extension.permit.r,
-          s: eip2612Extension.permit.s,
-          v: eip2612Extension.permit.v,
-        };
-
-        tx = await this.signer.writeContract({
-          address: x402ExactPermit2ProxyAddress,
-          abi: x402ExactPermit2ProxyABI,
-          functionName: "settleWith2612",
-          args: [
-            permit2612,
-            permit,
-            BigInt(requirements.amount),
-            getAddress(payer),
-            witness,
-            permit2Payload.signature,
-          ],
-        });
-      } else {
-        tx = await this.signer.writeContract({
-          address: x402ExactPermit2ProxyAddress,
-          abi: x402ExactPermit2ProxyABI,
-          functionName: "settle",
-          args: [
-            permit,
-            BigInt(requirements.amount),
-            getAddress(payer),
-            witness,
-            permit2Payload.signature,
-          ],
-        });
-      }
-
-      const receipt = await this.signer.waitForTransactionReceipt({ hash: tx });
-
-      if (receipt.status !== "success") {
-        return {
-          success: false,
-          errorReason: "invalid_transaction_state",
-          transaction: tx,
-          network: payload.accepted.network,
-          payer,
-        };
-      }
-
-      return {
-        success: true,
-        transaction: tx,
-        network: payload.accepted.network,
-        payer,
-      };
-    } catch (error) {
-      console.error("Failed to settle Permit2 transaction:", error);
-      return {
-        success: false,
-        errorReason: "transaction_failed",
-        transaction: "",
-        network: payload.accepted.network,
-        payer,
-      };
-    }
+    const eip3009Payload = rawPayload as ExactEIP3009Payload;
+    return settleEIP3009(
+      this.signer,
+      { deployERC4337WithEIP6492: this.config.deployERC4337WithEIP6492 },
+      payload,
+      requirements,
+      eip3009Payload,
+      () => verifyEIP3009(this.signer, payload, requirements, eip3009Payload),
+    );
   }
 }
